refactored the code to follow DRY
